existence check in almost all functions of the model to avoid sql errors

+++ model/model.js
@@ -27,7 +27,10 @@ db.prepare('CREATE TABLE IF NOT EXISTS projectEvents(projectId INTEGER REFERENCE
  * 
  *          FOR A USER
  * 
+ *
  */
+
+ /*No need of an existence check here, as the rout already check the username and email availability.*/
 exports.createUser = function(email, username, password, status) {
     return usersHandler.createUser(email, username, password, status);
 }
@@ -35,30 +38,33 @@ exports.createUser = function(email, username, password, status) {
 
 
 exports.updateUserPassword = function(email, password) {
+    if (!exists(email, "email", "users")) return null;
     return usersHandler.updateUserPassword(email, password);
 }
 
 exports.updateUserUsername = function(email, username) {
+    if (!exists(email, "email", "users")) return null;
     return usersHandler.updateUserUsername(email, username);
 }
 
 
 exports.deleteUser = function(email) {
+    if (!exists(email, "email", "users")) return null;
     return usersHandler.deleteUser(email);
 }
 
 exports.getUserStatus = function(userEmail) {
-    if (!sqlCheck("users", "email", userEmail)) return null;
+    if (!exists(userEmail, "email", "users")) return null;
     return usersHandler.getUserStatus(userEmail);
 }
 
 exports.getUserId = function(userUsername) {
-    if(!sqlCheck("users", "username", userUsername)) return null;
-    return usersHandler.getUserId(username);
+    if (!exists(userUsername, "username", "users")) return null;
+    return usersHandler.getUserId(userUsername);
 }
 
 exports.getUsername = function(userId) {
-    if (!sqlCheck("users", "email", userId)) return null;
+    if (!exists(userId, "email", "users")) return null;
     return usersHandler.getUsername(userId);
 }
 
@@ -76,11 +82,12 @@ exports.getProjects = function(username) {
  */
 
  exports.isTheRightPassword = function(email, userPassword) {
-     return usersHandler.isTheRightPassword(email, userPassword);
+    if(! exists(userEmail, "email", "users")) return null;
+    return usersHandler.isTheRightPassword(email, userPassword);
  }
 
  exports.credentialsAreFree = function(email, username) {
-     return usersHandler.credentialsAreFree(email, username);
+    return usersHandler.credentialsAreFree(email, username);
  }
 
 
@@ -91,6 +98,7 @@ exports.getProjects = function(username) {
  */
 
 exports.updateUserStatus = function (userEmail, newStatus) {
+    if(! exists(userEmail, "email", "users")) return null;
     return administratorsTools.updateUserStatus(userEmail, newStatus);
 }
 
@@ -107,14 +115,17 @@ exports.getUsersList = function() {
 
 
 exports.createProject = function(title, description, categories, creator, date, keywords) {
+    if(exists(projectId, "projectId", "projects")) return null;
     return projectsHandler.createProject(title, description, categories, creator, date, keywords);
 }
 
 exports.updateProject = function(projectId, title, description, creator) {
+    if(! exists(projectId, "projectId", "projects")) return null;
     return projectsHandler.updateProject(projectId, title, description, creator);
 }
 
 exports.deleteProject = function(projectId) {
+    if(! exists(projectId, "projectId", "projects")) return null;
     return projectsHandler.deleteProject(projectId);
 }
 
@@ -124,19 +135,23 @@ exports.deleteProject = function(projectId) {
  * 
  */
 
-exports.addMember = function(projectId, user, status) {
-    return projectsHandler.addMember(projectId, user, status);
+exports.addMember = function(projectIdConcerned, userToAdd, status) {
+    if(exists(projectIdConcerned, userToAdd, "projectId", "user", "projectMembers")) return null;
+    return projectsHandler.addMember(projectIdConcerned, userToAdd, status);
 }
 
-exports.updateMemberStatus = function(projectId, user, status) {
-    return projectsHandler.updateMemberStatus(projectId, user, status);
+exports.updateMemberStatus = function(projectIdConcerned, userConcerned, newStatus) {
+    if(! exists(projectIdConcerned, userConcerned, "projectId", "user", "projectMembers")) return null;
+    return projectsHandler.updateMemberStatus(projectIdConcered, userConcerned, newStatus);
 }
 
-exports.removeMember = function(projectId, user) {
-    return projectsHandler.removeMember(projectId, user);
+exports.removeMember = function(projectId, userToRemove) {
+    if(! exists(projectId, userToRemove, "projectId", "user", "projectMembers")) return null;
+    return projectsHandler.removeMember(projectId, userToRemove);
 }
 
 exports.getMembers = function(projectId) {
+    if(! exists(projectId, "projectId", "projectMembers")) return null;
     return projectsHandler.getMembers(projectId);
 }
 
@@ -146,12 +161,14 @@ exports.getMembers = function(projectId) {
  * 
  */
 
-exports.addKeyword = function(projectId, keyword) {
-    return keywordsHandler.addKeyword(projectId, keyword);
+exports.addKeyword = function(projectId, keywordToAdd) {
+    if(exists(projectId, keywordToAdd, "projectId", "keyword", "projectKeywords")) return null;
+    return keywordsHandler.addKeyword(projectId, keywordToAdd);
 }
 
-exports.removeKeyword = function(projectId, keyword) {
-    return keywordsHandler.removeKeyword(projectId, keyword);
+exports.removeKeyword = function(projectId, keywordToRemove) {
+    if(! exists(projectId, keywordToRemove, "projectId", "keyword", "projectKeywords")) return null;
+    return keywordsHandler.removeKeyword(projectId, keywordToRemove);
 }
 
 /***
@@ -160,15 +177,23 @@ exports.removeKeyword = function(projectId, keyword) {
  * 
  */
 
-exports.addEvent = function(projectId, event, date) {
-    return eventsHandler.addEvent(projectId, event, date);
+exports.addEvent = function(projectId, eventToAdd, dateOfEvent) {
+    if(exists(projectId, eventToAdd, "projectId", "event", "projectEvents")) return null;
+    return eventsHandler.addEvent(projectId, eventToAdd, dateOfEvent);
 }
 
 exports.updateEvent = function(projectId, previousEvent, newEvent) {
+    if(! exists(projectId, eventToRemove, "projectId", "event", "projectEvents")) return null;
     return eventsHandler.updateEvent(projectId, previousEvent, newEvent);
 }
 
-exports.removeEvent = function(projectId, event) {
+exports.changeEventDate = function(projectId, eventToChange, newDate) {
+    if(! exists(projectId, eventToChange, "projectId", "event", "projectEvents")) return null;
+    return eventsHandler.changeEventDate(projectId, eventToChange, newDate);
+}
+
+exports.removeEvent = function(projectId, eventToRemove) {
+    if(! exists(projectId, eventToRemove, "projectId", "event", "projectEvents")) return null;
     return eventsHandler.removeEvent(projectId, event);
 }
 
@@ -179,12 +204,15 @@ exports.removeEvent = function(projectId, event) {
  * 
  */
 
- exports.addCategory = function(projectId, category) {
-    return categoriesHandler.addCategory(projectId, category);
+ exports.addCategory = function(projectId, categoryToAdd) {
+    if(exists(projectId, categoryToAdd, "projectId", "category", "projectCategories")) return null;
+    return categoriesHandler.addCategory(projectId, categoryToAdd);
  }
 
- exports.removeCategory = function(projectId, category) {
-    return categoriesHandler.removeCategory(projectId, category);
+ exports.removeCategory = function(projectId, categoryToRemove) {
+    if(! exists(projectId, categoryToRemove, "projectId", "category", "projectCategories")) return null;
+
+    return categoriesHandler.removeCategory(projectId, categoryToRemove);
  }
 
 
@@ -198,6 +226,7 @@ exports.resetDatabase = function() {
     db.prepare('DROP TABLE projectEvents').run();
     db.prepare('DROP TABLE projectKeywords').run();
     db.prepare('DROP TABLE projectMembers').run();
+    db.prepare('DROP TABLE projectCategories').run();
     db.prepare('DROP TABLE projects').run();
     db.prepare('DROP TABLE users').run();
 
@@ -207,20 +236,27 @@ exports.resetDatabase = function() {
 
     db.prepare('CREATE TABLE projects (projectId INTEGER PRIMARY KEY AUTOINCREMENT, title VARCHAR2(60), description VARCHAR2(1000), creator VARCHAR2(30) REFERENCES users), date DATE').run();
 
-    db.prepare('CREATE TABLE projectMembers(projectId INTEGER REFERENCES projects ON DELETE CASCADE, user VARCHAR2(30) REFERENCES users ON DELETE CASCADE, status VARCHAR2(15), PRIMARY KEY(projectId, user))').run();
+    db.prepare('CREATE TABLE projectMembers (projectId INTEGER REFERENCES projects ON DELETE CASCADE, user VARCHAR2(30) REFERENCES users ON DELETE CASCADE, status VARCHAR2(15), PRIMARY KEY(projectId, user))').run();
 
-    db.prepare('CREATE TABLE projectKeyWords(projectId INTEGER REFERENCES projects ON DELETE CASCADE, keyword VARCHAR2(15), PRIMARY KEY(projectId, keyword))').run();
+    db.prepare('CREATE TABLE projectKeyWords (projectId INTEGER REFERENCES projects ON DELETE CASCADE, keyword VARCHAR2(15), PRIMARY KEY(projectId, keyword))').run();
 
     db;prepare('CREATE TABLE projectCategories (projectId INTEGER REFERENCES projects, category VARCHAR(20), PRIMARY KEY(projectId, category)').run();
 
-    db.prepare('CREATE TABLE projectEvents(projectId INTEGER REFERENCES projects ON DELETE CASCADE, event VARCHAR2(500), date DATE, PRIMARY KEY(projectId, event))').run();
+    db.prepare('CREATE TABLE projectEvents (projectId INTEGER REFERENCES projects ON DELETE CASCADE, event VARCHAR2(500), date DATE, PRIMARY KEY(projectId, event))').run();
 
     createUser('admin@admin.fr', 'Administrator', 'AZERTY', 'administrator');
 }
 
 
-var sqlCheck = function(table, field, content) {
+var exists = function(content, field, table) {
     content = String(content);
     let check = db.prepare(`SELECT ${field} FROM ${table} WHERE ${field}=?`).get([content]);
+    return check !== undefined;
+}
+
+var exists = function(content1, content2, field1, field2, table) {
+    content1 = String(content1);
+    content2 = String(content2);
+    let check = db.prepare(`SELECT ${field1}, ${field2} FROM ${table} WHERE ${field1}=? AND ${field2}=?`).get([content1, content2]);
     return check !== undefined;
 }
